noscope/requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
cryptography==3.0
decorator==4.4.2
discord.py==1.4.1
idna==2.9
mongoengine==0.20.0
multidict==4.7.5
pycparser==2.20
pymongo==3.11.0
python-dotenv==0.13.0
python-telegram-bot==12.8
PyYAML==5.3.1
redis==3.5.3
requests==2.24.0
rq==1.5.0
schedule==0.6.0
six==1.15.0
tornado==6.0.4
urllib3==1.25.10
websockets==8.1
yarl==1.4.2
